+++ bazel/core.bzl
@@ -184,12 +184,6 @@ MLN_CORE_SOURCE = [
     "src/mbgl/math/log2.cpp",
     "src/mbgl/platform/settings.cpp",
     "src/mbgl/programs/attributes.hpp",
-    "src/mbgl/programs/clipping_mask_program.hpp",
-    "src/mbgl/programs/program.hpp",
-    "src/mbgl/programs/program_parameters.cpp",
-    "src/mbgl/programs/program_parameters.hpp",
-    "src/mbgl/programs/programs.cpp",
-    "src/mbgl/programs/programs.hpp",
     "src/mbgl/programs/segment.hpp",
     "src/mbgl/programs/uniforms.hpp",
     "src/mbgl/renderer/backend_scope.cpp",
@@ -299,6 +293,7 @@ MLN_CORE_SOURCE = [
     "src/mbgl/renderer/transition_parameters.hpp",
     "src/mbgl/renderer/update_parameters.hpp",
     "src/mbgl/renderer/upload_parameters.hpp",
+    "src/mbgl/shaders/program_parameters.cpp",
     "src/mbgl/sprite/sprite_loader.cpp",
     "src/mbgl/sprite/sprite_loader.hpp",
     "src/mbgl/sprite/sprite_loader_observer.hpp",
@@ -650,6 +645,7 @@ MLN_CORE_HEADERS = [
     "include/mbgl/renderer/renderer_frontend.hpp",
     "include/mbgl/renderer/renderer_observer.hpp",
     "include/mbgl/renderer/renderer_state.hpp",
+    "include/mbgl/shaders/program_parameters.hpp",
     "include/mbgl/storage/database_file_source.hpp",
     "include/mbgl/storage/file_source.hpp",
     "include/mbgl/storage/file_source_manager.hpp",
@@ -854,6 +850,10 @@ MLN_OPENGL_SOURCE = [
     "src/mbgl/gl/vertex_buffer_resource.cpp",
     "src/mbgl/gl/vertex_buffer_resource.hpp",
     "src/mbgl/platform/gl_functions.cpp",
+    "src/mbgl/shaders/gl/legacy/clipping_mask_program.hpp",
+    "src/mbgl/shaders/gl/legacy/program.hpp",
+    "src/mbgl/shaders/gl/legacy/programs.cpp",
+    "src/mbgl/shaders/gl/legacy/programs.hpp",
 ]
 
 MLN_OPENGL_HEADERS = [
