+++ src/mbgl/renderer/render_static_data.hpp
@@ -4,7 +4,6 @@
 #include <mbgl/gfx/index_buffer.hpp>
 #include <mbgl/gfx/renderbuffer.hpp>
 #include <mbgl/gfx/shader_registry.hpp>
-#include <mbgl/programs/programs.hpp>
 #include <mbgl/renderer/buckets/heatmap_bucket.hpp>
 #include <mbgl/renderer/buckets/raster_bucket.hpp>
 
@@ -49,10 +48,6 @@ public:
     std::unique_ptr<gfx::ShaderRegistry> shaders;
 
     const SegmentVector clippingMaskSegments;
-
-#ifndef NDEBUG
-    Programs overdrawPrograms;
-#endif
 };
 
 } // namespace mbgl
