+++ src/mbgl/shaders/gl/legacy/programs.cpp
@@ -1,4 +1,5 @@
-#include <mbgl/programs/programs.hpp>
+#include <mbgl/shaders/gl/legacy/programs.hpp>
+#include <mbgl/shaders/gl/legacy/clipping_mask_program.hpp>
 #include <mbgl/util/logging.hpp>
 #include <exception>
