+++ src/mbgl/shaders/program_parameters.cpp
@@ -1,4 +1,4 @@
-#include <mbgl/programs/program_parameters.hpp>
+#include <mbgl/shaders/program_parameters.hpp>
 
 #include <mbgl/util/hash.hpp>
 #include <mbgl/util/string.hpp>
